perf: avoid redundant axisConfigType calculation
This may reduce ~5% of compile time
