+++ test/compile/axis/properties.test.ts
@@ -193,23 +193,27 @@ describe('compile/axis', () => {
     });
 
     it('should return the correct labelAngle from the axis definition', () => {
-      expect(240).toEqual(labelAngle(axisModel, axisModel.axis('y'), 'y', axisModel.typedFieldDef('y')));
+      expect(240).toEqual(labelAngle(axisModel, axisModel.axis('y'), 'y', axisModel.typedFieldDef('y'), []));
     });
 
     it('should return the correct labelAngle from the axis config definition', () => {
-      expect(140).toEqual(labelAngle(configModel, configModel.axis('y'), 'y', configModel.typedFieldDef('y')));
+      expect(140).toEqual(
+        labelAngle(configModel, configModel.axis('y'), 'y', configModel.typedFieldDef('y'), ['axis'])
+      );
     });
 
     it('should return the correct default labelAngle when not specified', () => {
-      expect(270).toEqual(labelAngle(defaultModel, defaultModel.axis('x'), 'x', defaultModel.typedFieldDef('x')));
+      expect(270).toEqual(labelAngle(defaultModel, defaultModel.axis('x'), 'x', defaultModel.typedFieldDef('x'), []));
     });
 
     it('should return the labelAngle declared in the axis when both the axis and axis config have labelAngle', () => {
-      expect(240).toEqual(labelAngle(bothModel, bothModel.axis('y'), 'y', bothModel.typedFieldDef('y')));
+      expect(240).toEqual(labelAngle(bothModel, bothModel.axis('y'), 'y', bothModel.typedFieldDef('y'), []));
     });
 
     it('should return undefined when there is no default and no specified labelAngle', () => {
-      expect(undefined).toEqual(labelAngle(neitherModel, neitherModel.axis('y'), 'y', neitherModel.typedFieldDef('y')));
+      expect(undefined).toEqual(
+        labelAngle(neitherModel, neitherModel.axis('y'), 'y', neitherModel.typedFieldDef('y'), [])
+      );
     });
   });
